+++ demo/Diffusion/demo_diffusion/pipeline/flux_pipeline.py
@@ -27,7 +27,7 @@ from typing import Any, List, Optional
 import numpy as np
 import tensorrt as trt
 import torch
-from cuda import cudart
+from cuda.bindings import runtime as cudart
 from diffusers.image_processor import VaeImageProcessor
 from huggingface_hub import snapshot_download