+++ demo/Diffusion/demo_diffusion/pipeline/model_memory_manager.py
@@ -16,7 +16,7 @@
 #
 
 import torch
-from cuda import cudart
+from cuda.bindings import runtime as cudart
 
 
 class ModelMemoryManager: