+++ demo/Diffusion/demo_diffusion/pipeline/stable_cascade_pipeline.py
@@ -20,7 +20,7 @@ import time
 
 import tensorrt as trt
 import torch
-from cuda import cudart
+from cuda.bindings import runtime as cudart
 from diffusers import DDPMWuerstchenScheduler
 
 from demo_diffusion.model import (