+++ demo/Diffusion/demo_img2img.py
@@ -18,7 +18,7 @@
 import argparse
 
 import PIL
-from cuda import cudart
+from cuda.bindings import runtime as cudart
 from PIL import Image
 
 from demo_diffusion import dd_argparse