+++ demo/Diffusion/demo_inpaint.py
@@ -17,7 +17,7 @@
 
 import argparse
 
-from cuda import cudart
+from cuda.bindings import runtime as cudart
 from PIL import Image
 
 from demo_diffusion import dd_argparse