+++ demo/Diffusion/demo_stable_cascade.py
@@ -19,7 +19,7 @@ import argparse
 import os
 
 import torch
-from cuda import cudart
+from cuda.bindings import runtime as cudart
 
 from demo_diffusion import dd_argparse
 from demo_diffusion import pipeline as pipeline_module